Allow access to sensor_ctl_socket as apk_data_file.
/data/app/sensor_ctl_socket should be labeled as sensors_socket
per the type_transition rule in sensors.te, but there are reports
of it being left in apk_data_file, yielding denials.

Allow access to apk_data_file:sock_file wherever we allow
access to sensors_socket (this is sometimes hidden by
unix_socket_* macros) and also add an auditallow rule so
that it will still be logged (as an avc: granted message).

Change-Id: Ibc7501860350e6bc4264e75581e0076ff566202a
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>