Allow access to sensor_ctl_socket under old type.
On an upgrade, the socket file will have the old type, so we need
to allow removing the old socket file so that it can be re-created
by sensors.  We also allow write access by camera so that the
attempt to connect prior to re-creation of the socket in the
new type does not fail with EACCES and trigger a failure in the camera
service.

Change-Id: Ia0c893200754e9d92a69c7d2afce7c976566e68a
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>